Refactor import paths for embedding functions and rename injection module
